requirements.txt
torch
gymnasium
numpy
pandas
matplotlib
akshare
dashscope
python-dotenv
